Extract participants list from the room inspector
